refactor: formalize recon engine with industrial gates
